+++ src/main/java/appeng/client/gui/implementations/GuiCraftingCPU.java
@@ -1,13 +1,3 @@
-/*
- * This file is part of Applied Energistics 2. Copyright (c) 2013 - 2014, AlgorithmX2, All rights reserved. Applied
- * Energistics 2 is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General
- * Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any
- * later version. Applied Energistics 2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General
- * Public License for more details. You should have received a copy of the GNU Lesser General Public License along with
- * Applied Energistics 2. If not, see <http://www.gnu.org/licenses/lgpl>.
- */
-
 package appeng.client.gui.implementations;
 
 import java.io.IOException;
@@ -15,7 +5,7 @@ import java.text.NumberFormat;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.Iterator;
+import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -26,8 +16,6 @@ import net.minecraft.client.renderer.RenderHelper;
 import net.minecraft.entity.player.InventoryPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.nbt.NBTTagList;
-import net.minecraftforge.common.util.Constants.NBT;
 
 import org.apache.commons.lang3.time.DurationFormatUtils;
 import org.lwjgl.opengl.GL11;
@@ -35,7 +23,6 @@ import org.lwjgl.opengl.GL12;
 
 import com.google.common.base.Joiner;
 
-import appeng.api.AEApi;
 import appeng.api.config.CraftingAllow;
 import appeng.api.config.Settings;
 import appeng.api.config.SortDir;
@@ -44,7 +31,6 @@ import appeng.api.config.ViewItems;
 import appeng.api.config.YesNo;
 import appeng.api.storage.data.IAEItemStack;
 import appeng.api.storage.data.IAEStack;
-import appeng.api.storage.data.IItemList;
 import appeng.api.util.NamedDimensionalCoord;
 import appeng.client.gui.AEBaseGui;
 import appeng.client.gui.IGuiTooltipHandler;
@@ -55,8 +41,8 @@ import appeng.client.gui.widgets.ISortSource;
 import appeng.client.gui.widgets.ITooltip;
 import appeng.client.gui.widgets.MEGuiTextField;
 import appeng.client.render.highlighter.BlockPosHighlighter;
-import appeng.container.AEBaseContainer;
 import appeng.container.implementations.ContainerCraftingCPU;
+import appeng.container.implementations.CraftingCpuEntry;
 import appeng.core.AEConfig;
 import appeng.core.AELog;
 import appeng.core.localization.ButtonToolTips;
@@ -66,7 +52,6 @@ import appeng.core.localization.Localization;
 import appeng.core.localization.PlayerMessages;
 import appeng.core.sync.network.NetworkHandler;
 import appeng.core.sync.packets.PacketCraftingItemInterface;
-import appeng.core.sync.packets.PacketCraftingRemainingOperations;
 import appeng.core.sync.packets.PacketInventoryAction;
 import appeng.core.sync.packets.PacketValueConfig;
 import appeng.helpers.InventoryAction;
@@ -79,12 +64,11 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
     protected static final int GUI_HEIGHT = 184;
     protected static final int GUI_WIDTH = 238;
 
-    protected static final int TEXTURE_BELOW_TOP_ROW_Y = 41;
-    protected static final int TEXTURE_ABOVE_BOTTOM_ROW_Y = 51;
     protected static final int DISPLAYED_ROWS = 6;
-
     protected static final int SECTION_LENGTH = 67;
     protected static final int SECTION_HEIGHT = 23;
+    protected static final int TEXTURE_BELOW_TOP_ROW_Y = 41;
+    protected static final int TEXTURE_ABOVE_BOTTOM_ROW_Y = 51;
 
     protected static final int SCROLLBAR_TOP = 19;
     protected static final int SCROLLBAR_LEFT = 218;
@@ -111,64 +95,46 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
     private static final int ICON_LOCK_MODE = 133;
     private static final int ICON_BLOCK_MODE = 134;
 
-    private final ContainerCraftingCPU craftingCpu;
-
-    protected IItemList<IAEStack<?>> storage = AEApi.instance().storage().createAEStackList();
-    protected IItemList<IAEStack<?>> active = AEApi.instance().storage().createAEStackList();
-    protected IItemList<IAEStack<?>> pending = AEApi.instance().storage().createAEStackList();
+    private final ContainerCraftingCPU container;
+    private final CraftingCpuVisualState visualState = new CraftingCpuVisualState();
+    private final RemainingOperationsTooltip remainingOperationsTooltip = new RemainingOperationsTooltip();
 
     protected int rows = DISPLAYED_ROWS;
 
-    private class RemainingOperations implements ITooltip {
+    private GuiButton cancel;
+    private GuiAeButton suspend;
+    private GuiImgButton toggleHideStored;
+    private GuiImgButton changeAllow;
+    private MEGuiTextField searchField;
+    private boolean hideStored;
+    private int remainingOperations;
+    private int hoveredVisibleIndex = -1;
+    private IAEStack<?> hoveredStack;
+    private NBTTagCompound hoveredInterfaceLocations;
 
-        private long refreshTick = System.currentTimeMillis();
-        private long lastWorkingTick = 0;
+    public GuiCraftingCPU(final InventoryPlayer inventoryPlayer, final Object target) {
+        this(new ContainerCraftingCPU(inventoryPlayer, target));
+    }
 
-        private int remainingOperations = 0;
-        private NBTTagCompound itemScheduledReasons = new NBTTagCompound();
+    protected GuiCraftingCPU(final ContainerCraftingCPU container) {
+        super(container);
+        this.container = container;
+        this.ySize = GUI_HEIGHT;
+        this.xSize = GUI_WIDTH;
+        this.hideStored = AEConfig.instance.getConfigManager().getSetting(Settings.HIDE_STORED) == YesNo.YES;
+        this.setScrollBar(new GuiScrollbar());
+    }
 
-        public long getLastWorkingTick() {
-            return lastWorkingTick;
-        }
+    public void clearItems() {
+        this.visualState.clear();
+        this.remainingOperations = 0;
+        this.hoveredVisibleIndex = -1;
+        this.hoveredStack = null;
+        this.hoveredInterfaceLocations = null;
+        this.rebuildFilteredEntries();
+    }
 
-        public long getRefreshTick() {
-            return refreshTick;
-        }
-
-        public void setLastWorkingTick(long lastWorkingTick) {
-            this.lastWorkingTick = lastWorkingTick;
-        }
-
-        public void setRefreshTick(long refreshTick) {
-            this.refreshTick = refreshTick;
-        }
-
-        public void setItemScheduledReasons(NBTTagCompound reasons) {
-            this.itemScheduledReasons = reasons == null ? new NBTTagCompound() : reasons;
-        }
-
-        public ScheduledReason getScheduledReason(IAEStack<?> stack) {
-            if (stack == null || this.itemScheduledReasons.hasNoTags()) {
-                return null;
-            }
-
-            final IAEStack<?> keyStack = stack.copy();
-            keyStack.setStackSize(1);
-            final String stackType = keyStack.getStackType().getId();
-            final int stackHash = keyStack.hashCode();
-
-            final NBTTagList entries = this.itemScheduledReasons.getTagList("Entries", NBT.TAG_COMPOUND);
-            for (int i = 0; i < entries.tagCount(); i++) {
-                final NBTTagCompound entry = entries.getCompoundTagAt(i);
-                if (stackType.equals(entry.getString("Type")) && stackHash == entry.getInteger("Hash")) {
-                    final int ordinal = entry.getInteger("Reason");
-                    if (ordinal >= 0 && ordinal < ScheduledReason.VALUES.length) {
-                        return ScheduledReason.VALUES[ordinal];
-                    }
-                }
-            }
-            return null;
-        }
+    private final class RemainingOperationsTooltip implements ITooltip {
 
         @Override
         public String getMessage() {
@@ -177,7 +143,7 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
 
         @Override
         public int xPos() {
-            return guiLeft + TITLE_LEFT_OFFSET + 200 - this.getStringWidth();
+            return guiLeft + TITLE_LEFT_OFFSET + 200 - this.getWidth();
         }
 
         @Override
@@ -187,7 +153,7 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
 
         @Override
         public int getWidth() {
-            return this.getStringWidth();
+            return fontRendererObj.getStringWidth(String.valueOf(remainingOperations));
         }
 
         @Override
@@ -199,129 +165,87 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
         public boolean isVisible() {
             return true;
         }
-
-        public void setRemainingOperations(int remainingOperations) {
-            this.remainingOperations = remainingOperations;
-        }
-
-        public int getRemainingOperations() {
-            return this.remainingOperations;
-        }
-
-        public int getStringWidth() {
-            return fontRendererObj.getStringWidth(String.valueOf(this.remainingOperations));
-        }
-    }
-
-    protected List<IAEStack<?>> visual = new ArrayList<>();
-    private List<IAEStack<?>> filteredVisual = new ArrayList<>();
-    private GuiButton cancel;
-    private GuiAeButton suspend;
-    protected List<IAEStack<?>> visualHiddenStored = new ArrayList<>();
-    protected GuiImgButton toggleHideStored;
-    protected boolean hideStored;
-    private int tooltip = -1;
-    private final RemainingOperations remainingOperations = new RemainingOperations();
-    private IAEStack<?> hoveredStack;
-    private NBTTagCompound hoveredStackNbt;
-    private GuiImgButton changeAllow;
-    private MEGuiTextField searchField;
-
-    public GuiCraftingCPU(final InventoryPlayer inventoryPlayer, final Object te) {
-        this(new ContainerCraftingCPU(inventoryPlayer, te));
-    }
-
-    protected GuiCraftingCPU(final ContainerCraftingCPU container) {
-        super(container);
-        this.craftingCpu = container;
-        this.ySize = GUI_HEIGHT;
-        this.xSize = GUI_WIDTH;
-        this.hideStored = AEConfig.instance.getConfigManager().getSetting(Settings.HIDE_STORED) == YesNo.YES;
-
-        final GuiScrollbar scrollbar = new GuiScrollbar();
-        this.setScrollBar(scrollbar);
-    }
-
-    public void clearItems() {
-        this.storage = AEApi.instance().storage().createAEStackList();
-        this.active = AEApi.instance().storage().createAEStackList();
-        this.pending = AEApi.instance().storage().createAEStackList();
-        this.visual = new ArrayList<>();
-        this.filteredVisual = new ArrayList<>();
-        this.visualHiddenStored = new ArrayList<>();
     }
 
     @Override
     protected void actionPerformed(final GuiButton btn) {
         super.actionPerformed(btn);
+
         if (this.cancel == btn) {
-            try {
-                NetworkHandler.instance.sendToServer(new PacketValueConfig("TileCrafting.Cancel", "Cancel"));
-            } catch (final IOException e) {
-                AELog.debug(e);
-            }
-        } else if (this.suspend == btn) {
-            try {
-                NetworkHandler.instance.sendToServer(new PacketValueConfig("TileCrafting.Suspend", "Suspend"));
-            } catch (final IOException e) {
-                AELog.debug(e);
-            }
-        } else if (this.toggleHideStored == btn) {
-            this.hideStored ^= true;
-            AEConfig.instance.getConfigManager().putSetting(Settings.HIDE_STORED, hideStored ? YesNo.YES : YesNo.NO);
-            this.toggleHideStored.set(hideStored ? YesNo.YES : YesNo.NO);
-            hideStoredSorting();
-            this.setScrollBar();
-            updateFilteredList();
-        } else if (btn == this.changeAllow) {
-            String msg = String.valueOf(((CraftingAllow) this.changeAllow.getCurrentValue()).ordinal());
-            try {
-                NetworkHandler.instance.sendToServer(new PacketValueConfig("TileCrafting.Allow", msg));
-            } catch (final IOException e) {
-                AELog.debug(e);
-            }
+            this.sendConfigPacket("TileCrafting.Cancel", "Cancel");
+            return;
         }
 
+        if (this.suspend == btn) {
+            this.sendConfigPacket("TileCrafting.Suspend", "Suspend");
+            return;
+        }
+
+        if (this.toggleHideStored == btn) {
+            this.hideStored = !this.hideStored;
+            AEConfig.instance.getConfigManager()
+                    .putSetting(Settings.HIDE_STORED, this.hideStored ? YesNo.YES : YesNo.NO);
+            this.toggleHideStored.set(this.hideStored ? YesNo.YES : YesNo.NO);
+            this.rebuildFilteredEntries();
+            return;
+        }
+
+        if (btn == this.changeAllow) {
+            final String msg = String.valueOf(this.changeAllow.getCurrentValue().ordinal());
+            this.sendConfigPacket("TileCrafting.Allow", msg);
+        }
+    }
+
+    private void sendConfigPacket(final String key, final String value) {
+        try {
+            NetworkHandler.instance.sendToServer(new PacketValueConfig(key, value));
+        } catch (final IOException e) {
+            AELog.debug(e);
+        }
     }
 
     @Override
     protected void mouseClicked(final int xCoord, final int yCoord, final int btn) {
-        if (isShiftKeyDown() && this.hoveredStackNbt != null) {
-            // when using the highlight feature in the crafting GUI we want to show all the interfaces
-            // that currently received items so the player can see if the items are processed properly
-            List<NamedDimensionalCoord> ndcl = NamedDimensionalCoord.readAsListFromNBTNamed(hoveredStackNbt);
-
-            if (!ndcl.isEmpty()) {
-                Map<NamedDimensionalCoord, String[]> ndcm = new HashMap<>();
-                for (NamedDimensionalCoord ndc : ndcl) {
-                    ndcm.put(
-                            ndc,
-                            new String[] { PlayerMessages.MachineHighlightedNamed.getUnlocalized(),
-                                    PlayerMessages.MachineInOtherDimNamed.getUnlocalized() });
-                }
-                BlockPosHighlighter.highlightNamedBlocks(
-                        mc.thePlayer,
-                        ndcm,
-                        ((Localization) () -> "tile.appliedenergistics2.BlockInterface.name").getLocal());
-                mc.thePlayer.closeScreen();
-            }
+        if (isShiftKeyDown() && this.hoveredInterfaceLocations != null) {
+            this.highlightHoveredInterfaces();
         } else if (this.hoveredStack != null && btn == 2) {
-            ((AEBaseContainer) inventorySlots).setTargetStack(this.hoveredStack);
-            final PacketInventoryAction p = new PacketInventoryAction(
+            container.setTargetStack(this.hoveredStack);
+            final PacketInventoryAction packet = new PacketInventoryAction(
                     InventoryAction.AUTO_CRAFT,
-                    inventorySlots.inventorySlots.size(),
+                    this.inventorySlots.inventorySlots.size(),
                     this.hoveredStack.getStackSize());
-            NetworkHandler.instance.sendToServer(p);
+            NetworkHandler.instance.sendToServer(packet);
         }
 
         super.mouseClicked(xCoord, yCoord, btn);
         this.searchField.mouseClicked(xCoord, yCoord, btn);
     }
 
+    private void highlightHoveredInterfaces() {
+        final List<NamedDimensionalCoord> blocks = NamedDimensionalCoord
+                .readAsListFromNBTNamed(this.hoveredInterfaceLocations);
+        if (blocks.isEmpty()) {
+            return;
+        }
+
+        final Map<NamedDimensionalCoord, String[]> messages = new HashMap<>();
+        for (final NamedDimensionalCoord block : blocks) {
+            messages.put(
+                    block,
+                    new String[] { PlayerMessages.MachineHighlightedNamed.getUnlocalized(),
+                            PlayerMessages.MachineInOtherDimNamed.getUnlocalized() });
+        }
+
+        BlockPosHighlighter.highlightNamedBlocks(
+                this.mc.thePlayer,
+                messages,
+                ((Localization) () -> "tile.appliedenergistics2.BlockInterface.name").getLocal());
+        this.mc.thePlayer.closeScreen();
+    }
+
     @Override
     public void initGui() {
         super.initGui();
-        this.setScrollBar();
         this.cancel = new GuiButton(
                 0,
                 this.guiLeft + CANCEL_LEFT_OFFSET,
@@ -342,103 +266,264 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
                 this.guiTop + this.ySize - 19,
                 Settings.HIDE_STORED,
                 AEConfig.instance.getConfigManager().getSetting(Settings.HIDE_STORED));
-        this.buttonList.add(this.toggleHideStored);
-        this.buttonList.add(this.cancel);
-        this.buttonList.add(this.suspend);
-
-        this.searchField = new MEGuiTextField(76, 12, "Search") {
-
-            @Override
-            public void onTextChange(String oldText) {
-                super.onTextChange(oldText);
-                updateFilteredList();
-            }
-        };
-        this.searchField.x = this.guiLeft + this.xSize - 101;
-        this.searchField.y = this.guiTop + 5;
-
         this.changeAllow = new GuiImgButton(
                 this.guiLeft - 20,
                 this.guiTop + 2,
                 Settings.CRAFTING_ALLOW,
                 CraftingAllow.ALLOW_ALL);
+
+        this.searchField = new MEGuiTextField(76, 12, "Search") {
+
+            @Override
+            public void onTextChange(final String oldText) {
+                super.onTextChange(oldText);
+                rebuildFilteredEntries();
+            }
+        };
+        this.searchField.x = this.guiLeft + this.xSize - 101;
+        this.searchField.y = this.guiTop + 5;
+
+        this.buttonList.add(this.toggleHideStored);
+        this.buttonList.add(this.cancel);
+        this.buttonList.add(this.suspend);
         this.buttonList.add(this.changeAllow);
+        this.rebuildFilteredEntries();
     }
 
-    private void setScrollBar() {
-        int size = filteredVisual.size();
-        this.getScrollBar().setTop(SCROLLBAR_TOP).setLeft(SCROLLBAR_LEFT).setHeight(SCROLLBAR_HEIGHT);
-        this.getScrollBar().setRange(0, (size + 2) / ITEMS_PER_ROW - rows, 1);
+    private void rebuildFilteredEntries() {
+        final String searchText = this.searchField == null ? "" : this.searchField.getText();
+        this.visualState.rebuildFilteredEntries(this.hideStored, searchText);
+        this.updateScrollBar();
+    }
+
+    protected void updateScrollBar() {
+        final int size = this.visualState.filteredSize();
+        this.getScrollBar().setTop(SCROLLBAR_TOP).setLeft(SCROLLBAR_LEFT).setHeight(this.getScrollBarHeight());
+        this.getScrollBar().setRange(0, (size + 2) / ITEMS_PER_ROW - this.rows, 1);
+    }
+
+    protected int getScrollBarHeight() {
+        return SCROLLBAR_HEIGHT;
+    }
+
+    protected boolean hasVisualEntries() {
+        return !this.visualState.isEmpty();
     }
 
     @Override
     public void drawScreen(final int mouseX, final int mouseY, final float btn) {
-        this.cancel.enabled = !this.visual.isEmpty();
-        this.suspend.visible = !this.visual.isEmpty();
+        this.cancel.enabled = this.hasVisualEntries();
+        this.suspend.visible = this.hasVisualEntries();
         this.updateSuspendButtonText();
-        this.changeAllow.set(CraftingAllow.values()[this.craftingCpu.allow]);
+        this.changeAllow.set(CraftingAllow.values()[this.container.allow]);
 
-        final int gx = (this.width - this.xSize) / 2;
-        final int gy = (this.height - this.ySize) / 2;
+        this.hoveredVisibleIndex = this.resolveHoveredIndex(mouseX, mouseY);
+        this.handleTooltip(mouseX, mouseY, this.remainingOperationsTooltip);
+        super.drawScreen(mouseX, mouseY, btn);
+    }
 
-        this.tooltip = -1;
+    private int resolveHoveredIndex(final int mouseX, final int mouseY) {
+        final int gridLeft = (this.width - this.xSize) / 2;
+        final int gridTop = (this.height - this.ySize) / 2;
+        int row = 0;
+        int column = 0;
 
-        final int offY = SECTION_HEIGHT;
-        int y = 0;
+        for (int slotIndex = 0; slotIndex <= ITEMS_PER_ROW * this.rows; slotIndex++) {
+            final int minX = gridLeft + ITEMSTACK_LEFT_OFFSET + column * SECTION_LENGTH;
+            final int minY = gridTop + ITEMSTACK_TOP_OFFSET + row * SECTION_HEIGHT;
+
+            if (minX < mouseX && minX + SECTION_LENGTH > mouseX && minY < mouseY && minY + SECTION_HEIGHT > mouseY) {
+                return slotIndex;
+            }
+
+            column++;
+            if (column == ITEMS_PER_ROW) {
+                row++;
+                column = 0;
+            }
+        }
+
+        return -1;
+    }
+
+    @Override
+    public void drawFG(final int offsetX, final int offsetY, final int mouseX, final int mouseY) {
+        this.drawHeader();
+        this.drawVisibleEntries();
+    }
+
+    private void drawHeader() {
+        String title = this.getGuiDisplayName(GuiText.CraftingStatus.getLocal());
+
+        if (this.container.getElapsedTime() > 0 && this.hasVisualEntries()) {
+            final long elapsedInMilliseconds = TimeUnit.MILLISECONDS
+                    .convert(this.container.getElapsedTime(), TimeUnit.NANOSECONDS);
+            final String elapsedTimeText = DurationFormatUtils
+                    .formatDuration(elapsedInMilliseconds, GuiText.ETAFormat.getLocal());
+            title = title.isEmpty() ? elapsedTimeText : title + " - " + elapsedTimeText;
+        }
+
+        this.fontRendererObj.drawString(
+                String.valueOf(this.remainingOperations),
+                TITLE_LEFT_OFFSET + 128 - this.remainingOperationsTooltip.getWidth(),
+                TITLE_TOP_OFFSET,
+                GuiColors.CraftingCPUTitle.getColor());
+        this.fontRendererObj
+                .drawString(title, TITLE_LEFT_OFFSET, TITLE_TOP_OFFSET, GuiColors.CraftingCPUTitle.getColor());
+    }
+
+    private void drawVisibleEntries() {
+        final int viewStart = this.getScrollBar().getCurrentScroll() * ITEMS_PER_ROW;
+        final int viewEnd = Math.min(viewStart + ITEMS_PER_ROW * this.rows, this.visualState.filteredSize());
+        final List<CraftingCpuEntry> filteredEntries = this.visualState.filteredEntries();
+        final ReadableNumberConverter converter = ReadableNumberConverter.INSTANCE;
+        final IAEStack<?> lastHoveredStack = this.hoveredStack;
+
+        this.hoveredStack = null;
+
         int x = 0;
-        for (int z = 0; z <= ITEMS_PER_ROW * rows; z++) {
-            final int minX = gx + ITEMSTACK_LEFT_OFFSET + x * SECTION_LENGTH;
-            final int minY = gy + ITEMSTACK_TOP_OFFSET + y * offY;
-
-            if (minX < mouseX && minX + SECTION_LENGTH > mouseX) {
-                if (minY < mouseY && minY + offY > mouseY) {
-                    this.tooltip = z;
-                    break;
-                }
+        int y = 0;
+        for (int index = viewStart; index < viewEnd; index++) {
+            final CraftingCpuEntry entry = filteredEntries.get(index);
+            final EntryTooltip tooltipData = this
+                    .drawEntry(entry, index - viewStart, x, y, converter, lastHoveredStack);
+            if (tooltipData != null) {
+                this.drawTooltip(tooltipData.x, tooltipData.y, tooltipData.message);
             }
 
             x++;
-
-            if (x == ITEMS_PER_ROW) {
+            if (x > 2) {
                 y++;
                 x = 0;
             }
         }
-
-        this.handleTooltip(mouseX, mouseY, remainingOperations);
-        super.drawScreen(mouseX, mouseY, btn);
     }
 
-    private void updateFilteredList() {
-        filteredVisual.clear();
-        String searchText = this.searchField.getText().toLowerCase();
+    private EntryTooltip drawEntry(final CraftingCpuEntry entry, final int visibleIndex, final int x, final int y,
+            final ReadableNumberConverter converter, final IAEStack<?> lastHoveredStack) {
+        final boolean active = entry.hasActiveAmount();
+        final boolean scheduled = entry.hasPendingAmount();
+        final int lines = (entry.hasStoredAmount() ? 1 : 0) + (active ? 1 : 0) + (scheduled ? 1 : 0);
+        final ScheduledReason scheduledReason = entry.getScheduledReason();
 
-        if (searchText.isEmpty()) {
-            filteredVisual.addAll(hideStored ? this.visualHiddenStored : this.visual);
-        } else {
-            List<IAEStack<?>> sourceList = hideStored ? this.visualHiddenStored : this.visual;
-            for (IAEStack<?> stack : sourceList) {
-                if (stack != null && stack.getDisplayName().toLowerCase().contains(searchText)) {
-                    filteredVisual.add(stack);
-                }
-            }
+        GL11.glPushMatrix();
+        GL11.glScaled(0.5, 0.5, 0.5);
+
+        if (AEConfig.instance.useColoredCraftingStatus && (active || scheduled)) {
+            final int startX = (x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET) * 2;
+            final int startY = ((y * SECTION_HEIGHT + ITEMSTACK_TOP_OFFSET) - 3) * 2;
+            final int endX = startX + (SECTION_LENGTH * 2);
+            final int endY = startY + (SECTION_HEIGHT * 2) - 2;
+            drawRect(startX, startY, endX, endY, this.getCraftingStateColor(scheduledReason, active));
         }
-    }
 
-    private void updateRemainingOperations() {
-        int interval = 1000;
-        if (this.remainingOperations.getRefreshTick() >= this.remainingOperations.getLastWorkingTick() + interval) {
-            try {
-                NetworkHandler.instance.sendToServer(new PacketCraftingRemainingOperations());
-            } catch (IOException ignored) {}
-            this.remainingOperations.setLastWorkingTick(this.remainingOperations.refreshTick);
-        } else {
-            this.remainingOperations.setRefreshTick(System.currentTimeMillis());
+        final LinkedList<String> lineList = new LinkedList<>();
+        final int negY = ((lines - 1) * 5) / 2;
+        int downY = 0;
+
+        if (entry.hasStoredAmount()) {
+            downY = this.drawAmountLine(
+                    GuiText.Stored.getLocal(),
+                    entry.getStoredAmount(),
+                    x,
+                    y,
+                    negY,
+                    downY,
+                    GuiColors.CraftingCPUStored.getColor(),
+                    visibleIndex,
+                    lineList,
+                    converter);
         }
+
+        if (entry.hasActiveAmount()) {
+            downY = this.drawAmountLine(
+                    GuiText.Crafting.getLocal(),
+                    entry.getActiveAmount(),
+                    x,
+                    y,
+                    negY,
+                    downY,
+                    GuiColors.CraftingCPUAmount.getColor(),
+                    visibleIndex,
+                    lineList,
+                    converter);
+        }
+
+        if (entry.hasPendingAmount()) {
+            this.drawAmountLine(
+                    GuiText.Scheduled.getLocal(),
+                    entry.getPendingAmount(),
+                    x,
+                    y,
+                    negY,
+                    downY,
+                    GuiColors.CraftingCPUScheduled.getColor(),
+                    visibleIndex,
+                    lineList,
+                    converter);
+        }
+
+        GL11.glPopMatrix();
+
+        final int iconX = x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET;
+        final int iconY = y * SECTION_HEIGHT + ITEMSTACK_TOP_OFFSET - 3;
+        this.drawScheduledReasonIcon(iconX, iconY, this.getScheduledReasonIconIndex(scheduledReason, active));
+
+        final int posX = x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET + SECTION_LENGTH - 19;
+        final int posY = y * SECTION_HEIGHT + ITEMSTACK_TOP_OFFSET;
+
+        EntryTooltip tooltipData = null;
+        if (this.hoveredVisibleIndex == visibleIndex) {
+            final IAEStack<?> hoveredVisualStack = entry.getVisualStack();
+            final boolean stackChanged = lastHoveredStack == null || !hoveredVisualStack.isSameType(lastHoveredStack);
+            final String tooltipMessage = this.buildTooltipMessage(hoveredVisualStack, lineList, stackChanged);
+            tooltipData = new EntryTooltip(
+                    x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET + SECTION_LENGTH - 8,
+                    y * SECTION_HEIGHT + ITEMSTACK_TOP_OFFSET + 10,
+                    tooltipMessage);
+            this.hoveredStack = hoveredVisualStack;
+        }
+
+        GL11.glPushAttrib(GL11.GL_ENABLE_BIT | GL11.GL_COLOR_BUFFER_BIT | GL11.GL_LIGHTING_BIT);
+        RenderHelper.enableGUIStandardItemLighting();
+        GL11.glEnable(GL12.GL_RESCALE_NORMAL);
+        GL11.glEnable(GL11.GL_DEPTH_TEST);
+        entry.getStack().drawInGui(this.mc, posX, posY);
+        GL11.glPopAttrib();
+
+        return tooltipData;
     }
 
-    private int getCraftingStateColor(ScheduledReason scheduledReason, boolean active) {
-        if (scheduledReason == null || scheduledReason == ScheduledReason.UNDEFINED) {
+    private int drawAmountLine(final String label, final long amount, final int x, final int y, final int negY,
+            final int downY, final int color, final int visibleIndex, final List<String> tooltipLines,
+            final ReadableNumberConverter converter) {
+        final String compactText = label + ": " + converter.toWideReadableForm(amount);
+        final int width = 4 + this.fontRendererObj.getStringWidth(compactText);
+        this.fontRendererObj.drawString(
+                compactText,
+                (int) ((x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET + SECTION_LENGTH - 19 - (width * 0.5)) * 2),
+                (y * SECTION_HEIGHT + ITEMSTACK_TOP_OFFSET + 6 - negY + downY) * 2,
+                color);
+
+        if (this.hoveredVisibleIndex == visibleIndex) {
+            tooltipLines.add(label + ": " + NumberFormat.getInstance().format(amount));
+        }
+
+        return downY + 5;
+    }
+
+    private String buildTooltipMessage(final IAEStack<?> refStack, final List<String> lineList,
+            final boolean stackChanged) {
+        String tooltipMessage = Platform.getItemDisplayName(refStack);
+        if (!lineList.isEmpty()) {
+            this.addItemTooltip(refStack, lineList, stackChanged);
+            tooltipMessage = tooltipMessage + '\n' + Joiner.on("\n").join(lineList);
+        }
+        return tooltipMessage;
+    }
+
+    private int getCraftingStateColor(final ScheduledReason scheduledReason, final boolean active) {
+        if (scheduledReason == ScheduledReason.UNDEFINED) {
             return active ? GuiColors.CraftingCPUActive.getColor() : GuiColors.CraftingCPUInactive.getColor();
         }
 
@@ -451,8 +536,8 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
         };
     }
 
-    private int getScheduledReasonIconIndex(ScheduledReason scheduledReason, boolean active) {
-        if (scheduledReason == null) {
+    private int getScheduledReasonIconIndex(final ScheduledReason scheduledReason, final boolean active) {
+        if (scheduledReason == ScheduledReason.UNDEFINED) {
             return -1;
         }
 
@@ -465,7 +550,7 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
         };
     }
 
-    private void drawScheduledReasonIcon(int x, int y, int iconIndex) {
+    private void drawScheduledReasonIcon(final int x, final int y, final int iconIndex) {
         if (iconIndex < 0) {
             return;
         }
@@ -486,219 +571,33 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
         GL11.glPopAttrib();
     }
 
-    @Override
-    public void drawFG(final int offsetX, final int offsetY, final int mouseX, final int mouseY) {
-        String title = this.getGuiDisplayName(GuiText.CraftingStatus.getLocal());
-
-        if (this.craftingCpu.getElapsedTime() > 0 && !this.visual.isEmpty()) {
-            final long elapsedInMilliseconds = TimeUnit.MILLISECONDS
-                    .convert(this.craftingCpu.getElapsedTime(), TimeUnit.NANOSECONDS);
-            final String elapsedTimeText = DurationFormatUtils
-                    .formatDuration(elapsedInMilliseconds, GuiText.ETAFormat.getLocal());
-
-            // If title is empty, don't show that ' - '
-            if (title.isEmpty()) {
-                title = elapsedTimeText;
-            } else {
-                title += " - " + elapsedTimeText;
-            }
+    protected void addItemTooltip(final IAEStack<?> refStack, final List<String> lineList, final boolean stackChanged) {
+        final ScheduledReason scheduledReason = this.visualState.findScheduledReason(refStack);
+        if (scheduledReason != ScheduledReason.UNDEFINED) {
+            lineList.add(scheduledReason.getLocal());
         }
-        updateRemainingOperations();
-        this.fontRendererObj.drawString(
-                String.valueOf(remainingOperations.getRemainingOperations()),
-                TITLE_LEFT_OFFSET + 128 - this.remainingOperations.getStringWidth(),
-                TITLE_TOP_OFFSET,
-                GuiColors.CraftingCPUTitle.getColor());
-
-        this.fontRendererObj
-                .drawString(title, TITLE_LEFT_OFFSET, TITLE_TOP_OFFSET, GuiColors.CraftingCPUTitle.getColor());
-
-        int x = 0;
-        int y = 0;
-        final int viewStart = this.getScrollBar().getCurrentScroll() * ITEMS_PER_ROW;
-        final int viewEnd = viewStart + ITEMS_PER_ROW * rows;
-
-        String dspToolTip = "";
-        final List<String> lineList = new LinkedList<>();
-        int toolPosX = 0;
-        int toolPosY = 0;
-
-        IAEStack<?> lastHoveredStack = this.hoveredStack;
-        this.hoveredStack = null;
-        final int offY = 23;
-
-        final ReadableNumberConverter converter = ReadableNumberConverter.INSTANCE;
-        List<IAEStack<?>> visualTemp = filteredVisual;
-        for (int z = viewStart; z < Math.min(viewEnd, visualTemp.size()); z++) {
-            final IAEStack<?> refStack = visualTemp.get(z); // repo.getReferenceItem( z );
-            if (refStack != null) {
-                GL11.glPushMatrix();
-                GL11.glScaled(0.5, 0.5, 0.5);
-
-                final IAEStack<?> stored = this.storage.findPrecise(refStack);
-                final IAEStack<?> activeStack = this.active.findPrecise(refStack);
-                final IAEStack<?> pendingStack = this.pending.findPrecise(refStack);
-
-                int lines = 0;
-
-                if (stored != null && stored.getStackSize() > 0) {
-                    lines++;
-                }
-                boolean active = false;
-                if (activeStack != null && activeStack.getStackSize() > 0) {
-                    lines++;
-                    active = true;
-                }
-                boolean scheduled = false;
-                if (pendingStack != null && pendingStack.getStackSize() > 0) {
-                    lines++;
-                    scheduled = true;
-                }
-
-                final ScheduledReason scheduledReason = this.remainingOperations.getScheduledReason(refStack);
-                final int scheduledIconIndex = this.getScheduledReasonIconIndex(scheduledReason, active);
-
-                if (AEConfig.instance.useColoredCraftingStatus && (active || scheduled)) {
-                    final int startX = (x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET) * 2;
-                    final int startY = ((y * offY + ITEMSTACK_TOP_OFFSET) - 3) * 2;
-                    final int endX = startX + (SECTION_LENGTH * 2);
-                    final int endY = startY + (offY * 2) - 2;
-
-                    int bgColor = this.getCraftingStateColor(scheduledReason, active);
-
-                    drawRect(startX, startY, endX, endY, bgColor);
-                }
-
-                final int negY = ((lines - 1) * 5) / 2;
-                int downY = 0;
-
-                if (stored != null && stored.getStackSize() > 0) {
-                    final String str = GuiText.Stored.getLocal() + ": "
-                            + converter.toWideReadableForm(stored.getStackSize());
-                    final int w = 4 + this.fontRendererObj.getStringWidth(str);
-                    this.fontRendererObj.drawString(
-                            str,
-                            (int) ((x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET + SECTION_LENGTH - 19 - (w * 0.5))
-                                    * 2),
-                            (y * offY + ITEMSTACK_TOP_OFFSET + 6 - negY + downY) * 2,
-                            GuiColors.CraftingCPUStored.getColor());
-
-                    if (this.tooltip == z - viewStart) {
-                        lineList.add(
-                                GuiText.Stored.getLocal() + ": "
-                                        + NumberFormat.getInstance().format(stored.getStackSize()));
-                    }
-
-                    downY += 5;
-                }
-
-                if (activeStack != null && activeStack.getStackSize() > 0) {
-                    final String str = GuiText.Crafting.getLocal() + ": "
-                            + converter.toWideReadableForm(activeStack.getStackSize());
-                    final int w = 4 + this.fontRendererObj.getStringWidth(str);
-
-                    this.fontRendererObj.drawString(
-                            str,
-                            (int) ((x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET + SECTION_LENGTH - 19 - (w * 0.5))
-                                    * 2),
-                            (y * offY + ITEMSTACK_TOP_OFFSET + 6 - negY + downY) * 2,
-                            GuiColors.CraftingCPUAmount.getColor());
-
-                    if (this.tooltip == z - viewStart) {
-                        this.hoveredStack = refStack;
-                        lineList.add(
-                                GuiText.Crafting.getLocal() + ": "
-                                        + NumberFormat.getInstance().format(activeStack.getStackSize()));
-                    }
-
-                    downY += 5;
-                }
-
-                if (pendingStack != null && pendingStack.getStackSize() > 0) {
-                    final String str = GuiText.Scheduled.getLocal() + ": "
-                            + converter.toWideReadableForm(pendingStack.getStackSize());
-                    final int w = 4 + this.fontRendererObj.getStringWidth(str);
-
-                    this.fontRendererObj.drawString(
-                            str,
-                            (int) ((x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET + SECTION_LENGTH - 19 - (w * 0.5))
-                                    * 2),
-                            (y * offY + ITEMSTACK_TOP_OFFSET + 6 - negY + downY) * 2,
-                            GuiColors.CraftingCPUScheduled.getColor());
-
-                    if (this.tooltip == z - viewStart) {
-                        lineList.add(
-                                GuiText.Scheduled.getLocal() + ": "
-                                        + NumberFormat.getInstance().format(pendingStack.getStackSize()));
-                    }
-                }
-
-                GL11.glPopMatrix();
-                final int posX = x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET + SECTION_LENGTH - 19;
-                final int posY = y * offY + ITEMSTACK_TOP_OFFSET;
-                final int iconX = x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET;
-                final int iconY = y * offY + ITEMSTACK_TOP_OFFSET - 3;
-
-                this.drawScheduledReasonIcon(iconX, iconY, scheduledIconIndex);
-
-                final IAEStack<?> is = refStack.copy();
-
-                if (this.tooltip == z - viewStart) {
-                    dspToolTip = Platform.getItemDisplayName(is);
-
-                    if (!lineList.isEmpty()) {
-                        boolean stackChanged = lastHoveredStack == null || !refStack.isSameType(lastHoveredStack);
-                        addItemTooltip(refStack, lineList, stackChanged);
-                        dspToolTip = dspToolTip + '\n' + Joiner.on("\n").join(lineList);
-                    }
-
-                    toolPosX = x * (1 + SECTION_LENGTH) + ITEMSTACK_LEFT_OFFSET + SECTION_LENGTH - 8;
-                    toolPosY = y * offY + ITEMSTACK_TOP_OFFSET;
-
-                    this.hoveredStack = is;
-                }
-
-                GL11.glPushAttrib(GL11.GL_ENABLE_BIT | GL11.GL_COLOR_BUFFER_BIT | GL11.GL_LIGHTING_BIT);
-                RenderHelper.enableGUIStandardItemLighting();
-                GL11.glEnable(GL12.GL_RESCALE_NORMAL);
-                GL11.glEnable(GL11.GL_DEPTH_TEST);
-                is.drawInGui(this.mc, posX, posY);
-                GL11.glPopAttrib();
-
-                x++;
-
-                if (x > 2) {
-                    y++;
-                    x = 0;
-                }
-            }
-        }
-
-        if (this.tooltip >= 0 && !dspToolTip.isEmpty()) {
-            this.drawTooltip(toolPosX, toolPosY + 10, dspToolTip);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    protected void addItemTooltip(IAEStack<?> refStack, List<String> lineList, boolean stackChanged) {
-        ScheduledReason sr = this.remainingOperations.getScheduledReason(refStack);
-        if (sr != null && sr != ScheduledReason.UNDEFINED) lineList.add(sr.getLocal());
 
         if (isShiftKeyDown()) {
-            final List<String> l = refStack instanceof IAEItemStack ais
-                    ? ais.getItemStack().getTooltip(this.mc.thePlayer, this.mc.gameSettings.advancedItemTooltips)
-                    : Collections.EMPTY_LIST;
-            if (!l.isEmpty()) l.remove(0);
-            lineList.addAll(l);
-            if (this.hoveredStackNbt == null || stackChanged) {
+            final List<String> tooltipLines = refStack instanceof IAEItemStack itemStack
+                    ? itemStack.getItemStack().getTooltip(this.mc.thePlayer, this.mc.gameSettings.advancedItemTooltips)
+                    : Collections.emptyList();
+            if (!tooltipLines.isEmpty()) {
+                tooltipLines.remove(0);
+            }
+            lineList.addAll(tooltipLines);
+
+            if (this.hoveredInterfaceLocations == null || stackChanged) {
                 try {
                     NetworkHandler.instance.sendToServer(new PacketCraftingItemInterface(refStack.copy()));
-                } catch (Exception ignored) {}
+                } catch (final Exception ignored) {}
             } else {
-                List<NamedDimensionalCoord> blocks = NamedDimensionalCoord.readAsListFromNBTNamed(this.hoveredStackNbt);
+                final List<NamedDimensionalCoord> blocks = NamedDimensionalCoord
+                        .readAsListFromNBTNamed(this.hoveredInterfaceLocations);
+                if (blocks.isEmpty()) {
+                    return;
+                }
 
-                if (blocks.isEmpty()) return;
-                for (NamedDimensionalCoord blockPos : blocks) {
+                for (final NamedDimensionalCoord blockPos : blocks) {
                     lineList.add(
                             String.format(
                                     "Dim:%s X:%s Y:%s Z:%s \"%s\"",
@@ -711,7 +610,7 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
                 lineList.add(GuiText.HoldShiftClick_HIGHLIGHT_INTERFACE.getLocal());
             }
         } else {
-            this.hoveredStackNbt = null;
+            this.hoveredInterfaceLocations = null;
             lineList.add(GuiText.HoldShiftForTooltip.getLocal());
         }
     }
@@ -720,7 +619,7 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
     public void drawBG(final int offsetX, final int offsetY, final int mouseX, final int mouseY) {
         this.bindTexture("guis/craftingcpu.png");
         this.drawTexturedModalRect(offsetX, offsetY, 0, 0, this.xSize, this.ySize);
-        drawSearch();
+        this.drawSearch();
     }
 
     public void drawSearch() {
@@ -731,157 +630,134 @@ public class GuiCraftingCPU extends AEBaseGui implements ISortSource, IGuiToolti
 
     @Override
     protected void keyTyped(final char character, final int key) {
-        if (!(this.searchField.textboxKeyTyped(character, key))) {
+        if (!this.searchField.textboxKeyTyped(character, key)) {
             super.keyTyped(character, key);
         }
     }
 
-    public void postUpdateTooltip(NBTTagCompound nbt) {
-        this.hoveredStackNbt = nbt;
+    public void postUpdateTooltip(final NBTTagCompound nbt) {
+        this.hoveredInterfaceLocations = nbt;
     }
 
-    public void postUpdateBatchReasons(NBTTagCompound nbt) {
-        this.remainingOperations.setItemScheduledReasons(nbt);
-    }
-
-    public void postUpdate(int remainingOperations) {
-        this.remainingOperations.setRemainingOperations(remainingOperations);
-    }
-
-    public void postUpdate(final List<IAEStack<?>> list, final byte ref) {
-        for (final IAEStack<?> stack : list) {
-            switch (ref) {
-                case 0 -> {
-                    this.handleInput(this.storage, stack);
-                }
-                case 1 -> {
-                    this.handleInput(this.active, stack);
-                }
-                case 2 -> {
-                    this.handleInput(this.pending, stack);
-                }
-            }
-        }
-
-        for (final IAEStack<?> stack : list) {
-            final long amt = this.getTotal(stack);
-
-            if (amt <= 0) {
-                this.deleteVisualStack(stack);
-            } else {
-                final IAEStack<?> is = this.findVisualStack(stack);
-                is.setStackSize(amt);
-            }
-        }
-
-        if (this.hideStored) this.hideStoredSorting();
-        updateFilteredList();
-        this.setScrollBar();
-    }
-
-    private void handleInput(final IItemList<IAEStack<?>> s, final IAEStack<?> l) {
-        IAEStack<?> a = s.findPrecise(l);
-
-        if (l.getStackSize() <= 0) {
-            if (a != null) {
-                a.reset();
-            }
-        } else {
-            if (a == null) {
-                s.add(l.copy());
-                a = s.findPrecise(l);
-            }
-
-            if (a != null) {
-                a.setStackSize(l.getStackSize());
-            }
-        }
-    }
-
-    private long getTotal(final IAEStack<?> is) {
-        final IAEStack<?> a = this.storage.findPrecise(is);
-        final IAEStack<?> b = this.active.findPrecise(is);
-        final IAEStack<?> c = this.pending.findPrecise(is);
-
-        long total = 0;
-
-        if (a != null) {
-            total += a.getStackSize();
-        }
-
-        if (b != null) {
-            total += b.getStackSize();
-        }
-
-        if (c != null) {
-            total += c.getStackSize();
-        }
-
-        return total;
-    }
-
-    private void deleteVisualStack(final IAEStack<?> l) {
-        final Iterator<IAEStack<?>> i = this.visual.iterator();
-
-        while (i.hasNext()) {
-            final IAEStack<?> o = i.next();
-            if (o.equals(l)) {
-                i.remove();
-                return;
-            }
-        }
-    }
-
-    private IAEStack<?> findVisualStack(final IAEStack<?> l) {
-        for (final IAEStack<?> o : this.visual) {
-            if (o.equals(l)) {
-                return o;
-            }
-        }
-
-        final IAEStack<?> stack = l.copy();
-        this.visual.add(stack);
-
-        return stack;
+    public void postVisualEntryUpdate(final CraftingCpuEntry[] updates, final boolean clearFirst,
+            final int remainingOperations) {
+        this.visualState.applyEntryUpdates(updates, clearFirst);
+        this.remainingOperations = remainingOperations;
+        this.rebuildFilteredEntries();
     }
 
     @Override
+    @SuppressWarnings("rawtypes")
     public Enum getSortBy() {
         return SortOrder.NAME;
     }
 
     @Override
+    @SuppressWarnings("rawtypes")
     public Enum getSortDir() {
         return SortDir.ASCENDING;
     }
 
     @Override
+    @SuppressWarnings("rawtypes")
     public Enum getSortDisplay() {
         return ViewItems.ALL;
     }
 
     @Override
     public ItemStack getHoveredStack() {
-        return hoveredStack != null ? this.hoveredStack.getItemStackForNEI() : null;
-    }
-
-    private void hideStoredSorting() {
-        this.visualHiddenStored = new ArrayList<>();
-        for (final IAEStack<?> refStack : this.visual) {
-            if (refStack != null) {
-                final IAEStack<?> activeStack = this.active.findPrecise(refStack);
-                final IAEStack<?> pendingStack = this.pending.findPrecise(refStack);
-                if ((activeStack != null && activeStack.getStackSize() > 0)
-                        || (pendingStack != null && pendingStack.getStackSize() > 0)) {
-                    this.visualHiddenStored.add(refStack);
-                }
-            }
-        }
+        return this.hoveredStack != null ? this.hoveredStack.getItemStackForNEI() : null;
     }
 
     private void updateSuspendButtonText() {
-        var suspended = this.craftingCpu.cachedSuspend;
+        final boolean suspended = this.container.cachedSuspend;
         this.suspend.displayString = suspended ? GuiText.Resume.getLocal() : GuiText.Suspend.getLocal();
         this.suspend.setTootipString(suspended ? ButtonToolTips.Resume.getLocal() : ButtonToolTips.Suspend.getLocal());
+    }
+
+    private static final class EntryTooltip {
+
+        private final int x;
+        private final int y;
+        private final String message;
+
+        private EntryTooltip(final int x, final int y, final String message) {
+            this.x = x;
+            this.y = y;
+            this.message = message;
+        }
+    }
+
+    private static final class CraftingCpuVisualState {
+
+        private final Map<IAEStack<?>, CraftingCpuEntry> entries = new LinkedHashMap<>();
+        private final List<CraftingCpuEntry> filteredEntries = new ArrayList<>();
+
+        public void clear() {
+            this.entries.clear();
+            this.filteredEntries.clear();
+        }
+
+        public void applyEntryUpdates(final CraftingCpuEntry[] updates, final boolean clearFirst) {
+            if (clearFirst) {
+                this.clear();
+            }
+
+            if (updates == null) {
+                return;
+            }
+
+            for (final CraftingCpuEntry update : updates) {
+                if (update == null || update.getStack() == null) {
+                    continue;
+                }
+
+                final IAEStack<?> key = CraftingCpuEntry.normalizeStack(update.getStack());
+                if (update.getTotalAmount() <= 0) {
+                    this.entries.remove(key);
+                    continue;
+                }
+
+                this.entries.put(key, update);
+            }
+        }
+
+        public void rebuildFilteredEntries(final boolean hideStored, final String searchText) {
+            final String normalizedSearchText = searchText == null ? "" : searchText.toLowerCase();
+            this.filteredEntries.clear();
+
+            for (final CraftingCpuEntry entry : this.entries.values()) {
+                if (hideStored && !entry.isVisibleWhenHideStoredEnabled()) {
+                    continue;
+                }
+
+                if (!entry.matchesSearch(normalizedSearchText)) {
+                    continue;
+                }
+
+                this.filteredEntries.add(entry);
+            }
+        }
+
+        public boolean isEmpty() {
+            return this.entries.isEmpty();
+        }
+
+        public int filteredSize() {
+            return this.filteredEntries.size();
+        }
+
+        public List<CraftingCpuEntry> filteredEntries() {
+            return this.filteredEntries;
+        }
+
+        public ScheduledReason findScheduledReason(final IAEStack<?> stack) {
+            if (stack == null) {
+                return ScheduledReason.UNDEFINED;
+            }
+
+            final CraftingCpuEntry entry = this.entries.get(CraftingCpuEntry.normalizeStack(stack));
+            return entry == null ? ScheduledReason.UNDEFINED : entry.getScheduledReason();
+        }
     }
 }
