+++ src/main/java/appeng/core/sync/AppEngPacketHandlerBase.java
@@ -28,7 +28,6 @@ import appeng.core.sync.packets.PacketCraftingCPUsUpdate;
 import appeng.core.sync.packets.PacketCraftingCompleteNotification;
 import appeng.core.sync.packets.PacketCraftingCpuVisualEntries;
 import appeng.core.sync.packets.PacketCraftingItemInterface;
-import appeng.core.sync.packets.PacketCraftingRemainingOperations;
 import appeng.core.sync.packets.PacketCraftingTreeData;
 import appeng.core.sync.packets.PacketGuiDataSync;
 import appeng.core.sync.packets.PacketHighlightBlockStorage;
@@ -126,7 +125,6 @@ public class AppEngPacketHandlerBase {
         PACKET_PATTERN_VALUE(PacketPatternValueSet.class),
         PACKET_PATTERN_MULTI(PacketPatternMultiSet.class),
         PACKET_CRAFTING_CPU_VISUAL_ENTRIES(PacketCraftingCpuVisualEntries.class),
-        PACKET_CRAFTING_REMAINING_OPERATIONS(PacketCraftingRemainingOperations.class),
         PACKET_CRAFTING_ITEM_INTERFACE(PacketCraftingItemInterface.class),
         PACKET_CRAFTING_TREE_DATA(PacketCraftingTreeData.class),
         PACKET_NEI_BOOKMARK(PacketNEIBookmark.class),
